Iterating on the track page
