+++ assets/js/gens.ts
@@ -3,17 +3,18 @@ import "./components/input_controls";
 import "./components/util/popup";
 import "./components/util/shadowbaseelement";
 import "./components/util/choice_select";
-import "./components/side_menu";
 import "./components/side_menu/settings_page";
+import "./components/side_menu/side_menu";
 import "./components/header_info";
 import "./components/util/marker";
+import "./components/side_menu/track_page"
 
 import { API } from "./state/api";
 import { TracksManager } from "./components/tracks_manager";
 import { InputControls } from "./components/input_controls";
 import { getRenderDataSource } from "./state/parse_data";
 import { CHROMOSOMES } from "./constants";
-import { SideMenu } from "./components/side_menu";
+import { SideMenu } from "./components/side_menu/side_menu";
 import { SettingsPage } from "./components/side_menu/settings_page";
 import { HeaderInfo } from "./components/header_info";
 import { GensSession } from "./state/session";
@@ -135,7 +136,7 @@ export async function initCanvases({
     sideMenu.showContent("Settings", [settingsPage]);
 
     if (!settingsPage.isInitialized) {
-      settingsPage.initialize(render);
+      settingsPage.initialize();
     }
   });
 }
